+++ core/synthesis/dimensional_synthesis/ds_dialog/thread.py
@@ -14,7 +14,11 @@ from typing import (
     Any,
 )
 from time import time
-from platform import system
+from platform import (
+    system,
+    release,
+    machine,
+)
 from psutil import virtual_memory
 import numpy
 import numpy.distutils.cpuinfo
@@ -88,7 +92,7 @@ class WorkerThread(QThread):
             'interrupted': str(last_gen) if self.is_stop else 'False',
             'settings': self.settings,
             'hardware_info': {
-                'os': f"{system()} {platform.release()} {platform.machine()}",
+                'os': f"{system()} {release()} {machine()}",
                 'memory': f"{virtual_memory().total / (1 << 30):.04f} GB",
                 'cpu': cpu.get("model name", cpu.get('ProcessorNameString', '')),
             },